move samples/javafx project to samples/kotlin/javafx
