+++ samples/kotlin/javafx/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/sample/kotlin/javafx/ui/SampleView.kt
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package io.matthewnelson.kmp.tor.sample.javafx.ui
+package io.matthewnelson.kmp.tor.sample.kotlin.javafx.ui
 
-import io.matthewnelson.kmp.tor.sample.javafx.util.Log
+import io.matthewnelson.kmp.tor.sample.kotlin.javafx.util.Log
 import tornadofx.*
 
 class SampleView: View() {
